Moved flatten_linelist
